refactor: extract sync settings domain
