+++ src/sync/domains/index.ts
@@ -1,6 +1,7 @@
 export * from "@/sync/domains/calendar";
 export * from "@/sync/domains/contacts";
 export * from "@/sync/domains/files";
+export * from "@/sync/domains/settings";
 export * from "@/sync/domains/songs";
 export * from "@/sync/domains/stickies";
 export * from "@/sync/domains/videos";
